+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 + Add two more chat filter categories: Powder Mining and Winter Gifts.
 + Add catacombs class level color to party finder.
 + Add wishing compass uses amount display.
++ Saves missing items from cancelled buy orders to clipboard for faster re-entry.
 
 ### Removals
 - Removed additional settings for the chat design, like channel prefix and skyblock level